requirements.txt
pandas==2.2.3
requests==2.32.3
boto3==1.36.26
psycopg2-binary==2.9.10
pyyaml==6.0.2
python-dotenv==1.0.1
